td-auth-master/requirements/local.txt
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.5.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-stripe==2.4.3
Django==2.2
django-allauth==0.44.0
django-cleanup==5.1.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-model-utils==4.1.1
django-redis==4.12.1
django-stubs==1.7.0
docutils==0.16
factory-boy==3.2.0
Faker==6.6.3
filelock==3.0.12
flake8==3.9.0
flake8-isort==4.0.0
hiredis==1.1.0
identify==2.2.1
idna==2.10
imagesize==1.2.0
importlib-metadata==3.8.1
importlib-resources==5.1.2
iniconfig==1.1.1
ipdb==0.13.7
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
jsonfield==3.1.0
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
pre-commit==2.11.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.1.0
sorl-thumbnail==12.7.0
Sphinx==3.5.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
stripe==2.56.0
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1

channels~=3.0.3

  


  
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==2.2.15
django-bootstrap-static==4.2.1
django-cropperjs==0.0.5
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
msgpack==1.0.0
numpy==1.19.1
opencv-python==4.4.0.40
optional-django==0.1.0
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
webencodings==0.5.1
zipp==3.1.0
zope.interface==5.1.0
